refactor(xai): keep RealtimeSession in realtime_model.py
Avoid splitting the plugin module in this PR; behavior is unchanged.
